requirements.txt
# Core frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.14.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Model training and optimization
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
# deepspeed>=0.12.0  # May require special setup on Windows

# Evaluation tools
rouge-score>=0.1.2
sacrebleu>=2.3.0
bert-score>=0.3.13
# bleurt>=0.0.2  # Not available on PyPI, install manually if needed

# Visualization and monitoring
wandb>=0.15.0
tensorboard>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utility libraries
tqdm>=4.65.0
PyYAML>=6.0
click>=8.1.0
requests>=2.31.0
huggingface-hub>=0.17.0

# Jupyter support (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Other utilities
psutil>=5.9.0
gpustat>=1.1.0
nvitop>=1.3.0